Move Offset type to Cornelis.Offsets; remove Data.Data deriving
