+++ cornelis.cabal
@@ -50,7 +50,6 @@ library
       ConstrainedClassMethods
       ConstraintKinds
       DataKinds
-      DeriveDataTypeable
       DeriveFoldable
       DeriveFunctor
       DeriveGeneric
@@ -133,7 +132,6 @@ executable cornelis
       ConstrainedClassMethods
       ConstraintKinds
       DataKinds
-      DeriveDataTypeable
       DeriveFoldable
       DeriveFunctor
       DeriveGeneric
@@ -221,7 +219,6 @@ test-suite test
       ConstrainedClassMethods
       ConstraintKinds
       DataKinds
-      DeriveDataTypeable
       DeriveFoldable
       DeriveFunctor
       DeriveGeneric
