+++ src/Cornelis/Types/Agda.hs
@@ -9,11 +9,8 @@ import           Control.Monad.Except (ExceptT, throwError)
 import           Control.Monad.State.Strict (StateT, runStateT, put, get)
 import           Control.Monad.Trans (lift)
 import           Control.Monad.Trans.Except (runExceptT)
-import           Data.Aeson (FromJSON)
-import           Data.Data
 import           Data.Foldable (toList)
 import           Data.Functor.Identity
-import           Data.Int
 import qualified Data.List as List
 import           Data.Maybe (listToMaybe)
 import           Data.Sequence
@@ -21,25 +18,8 @@ import qualified Data.Sequence as Seq
 import           Data.Text (Text)
 import           GHC.Generics
 import           GHC.Show (showSpace)
-import           Prettyprinter (Pretty)
 import           System.FilePath
-
-------------------------------------------------------------------------------
--- | Line numbers are always 1-indexed
-newtype LineNumber = LineNumber { getOneIndexedLineNumber :: Int32 }
-  deriving stock Data
-  deriving newtype (Eq, Ord, Show, Read, FromJSON, Pretty)
-
-data OffsetType = Line | File | OneIndexed
-
-newtype Offset (a :: OffsetType) = Offset Int32
-  deriving stock Data
-  deriving newtype (Eq, Ord, Show, Read, FromJSON, Pretty)
-
-type BufferOffset = Offset 'File
-type LineOffset = Offset 'Line
-type AgdaOffset = Offset 'OneIndexed
-
+import           Cornelis.Offsets
 
 data Rewrite =  AsIs | Instantiated | HeadNormal | Simplified | Normalised
     deriving (Show, Read, Eq, Ord, Enum, Bounded)
@@ -85,7 +65,7 @@ data Range' a
   = NoRange
   | Range !a (Seq IntervalWithoutFile)
   deriving
-    (Data, Eq, Ord, Functor, Foldable, Traversable, Generic)
+    (Eq, Ord, Functor, Foldable, Traversable, Generic)
 
 instance Show a =>
          Show (Range' a) where
@@ -104,7 +84,7 @@ instance Show a =>
                  showSpace (showsPrec 11 $ toList b2_a1hOm))))
 
 data Interval' a = Interval { iStart, iEnd :: !(Pos' a) }
-  deriving (Data, Eq, Ord, Functor, Foldable, Traversable, Generic)
+  deriving (Eq, Ord, Functor, Foldable, Traversable, Generic)
 
 instance Show a => Show (Interval' a) where
   showsPrec n (Interval s e) =
@@ -118,11 +98,11 @@ data Pos' a = Pos
   { p_line :: !LineNumber
   , p_col  :: !a
   }
-  deriving (Eq, Ord, Show, Data, Functor, Foldable, Traversable, Generic)
+  deriving (Eq, Ord, Show, Functor, Foldable, Traversable, Generic)
 
 
 newtype AbsolutePath = AbsolutePath { textPath :: String }
-  deriving (Eq, Ord, Data)
+  deriving (Eq, Ord)
 
 instance Show AbsolutePath where
   showsPrec n (AbsolutePath p) =
@@ -318,14 +298,14 @@ data HighlightingLevel
     -- ^ This includes both non-interactive highlighting and
     -- interactive highlighting of the expression that is currently
     -- being type-checked.
-    deriving (Eq, Ord, Show, Read, Data, Generic)
+    deriving (Eq, Ord, Show, Read, Generic)
 
 data HighlightingMethod
   = Direct
     -- ^ Via stdout.
   | Indirect
     -- ^ Both via files and via stdout.
-    deriving (Eq, Show, Read, Data, Generic)
+    deriving (Eq, Show, Read, Generic)
 
 
 data Remove
